feature(saas runner): Refactor main.go and extract duplicated code into common package
Renamed main.go and moved common functions and environment variables handling to the common package to enhance code modularity. This refactoring simplifies the main logic and prevents redundancy by centralizing shared functionalities.
